Move network utility to :dependencyManagement project
