+++ subprojects/dependency-management/src/test/groovy/org/gradle/api/internal/artifacts/repositories/transport/NetworkOperationBackOffAndRetryTest.groovy
@@ -14,10 +14,11 @@
  * limitations under the License.
  */
 
-package org.gradle.api.publish.maven.internal.publisher
+package org.gradle.api.internal.artifacts.repositories.transport
 
 import org.apache.http.conn.HttpHostConnectException
 import org.gradle.api.UncheckedIOException
+import org.gradle.api.internal.artifacts.repositories.transport.NetworkOperationBackOffAndRetry
 import org.gradle.internal.resource.transport.http.HttpErrorStatusCodeException
 import org.gradle.testing.internal.util.Specification
 import spock.lang.Unroll
